online-shopping-site-1024 a test API

+++ online-shopping-site-1024/user-service/src/main/java/org/frank/services/AddressService.java
@@ -4,5 +4,5 @@ import org.frank.models.AddressDO;
 
 public interface AddressService {
     
-    AddressDO detail(int id);
+    AddressDO detail(Long id);
 }
